Desktop/bias/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic==0.39.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
